improved performance slightly

+++ src/InputHandler.cpp
@@ -18,13 +18,13 @@ InputHandler::~InputHandler()
 
 bool InputHandler::update()
 {
-	for (unsigned short i = 0; i < sf::Keyboard::KeyCount; ++i)
+	for (unsigned int i = 0; i < sf::Keyboard::KeyCount; ++i)
 	{
 		previousKeyState[i] = currentKeyState[i];
 		currentKeyState[i] = sf::Keyboard::isKeyPressed(static_cast<sf::Keyboard::Key>(i));
 	}
 
-	for (unsigned short i = 0; i < sf::Mouse::ButtonCount; ++i)
+	for (unsigned int i = 0; i < sf::Mouse::ButtonCount; ++i)
 	{
 		previousButtonState[i] = currentButtonState[i];
 		currentButtonState[i] = sf::Mouse::isButtonPressed(static_cast<sf::Mouse::Button>(i));
